@@ -1,94 +1,57 @@
-import React, { Component } from 'react';
-import {
-  Route,
-  Switch
-} from 'react-router-dom';
+import React, {useEffect, useState} from 'react';
+import {BrowserRouter, Route, Routes} from 'react-router-dom';
 import AppHeader from '../common/AppHeader';
-import Home from '../home/Home';
-import Login from '../user/login/Login';
-import Signup from '../user/signup/Signup';
-import Profile from '../user/profile/Profile';
-import OAuth2RedirectHandler from '../user/oauth2/OAuth2RedirectHandler';
-import NotFound from '../common/NotFound';
 import LoadingIndicator from '../common/LoadingIndicator';
-import { getCurrentUser } from '../util/APIUtils';
-import { ACCESS_TOKEN } from '../constants';
-import PrivateRoute from '../common/PrivateRoute';
-import Alert from 'react-s-alert';
-import 'react-s-alert/dist/s-alert-default.css';
-import 'react-s-alert/dist/s-alert-css-effects/slide.css';
+import {ACCESS_TOKEN} from '../constants';
 import './App.css';
+import NotFound from "../common/NotFound";
+import OAuth2RedirectHandler from "../user/oauth2/OAuth2RedirectHandler";
+import Home from "../home/Home";
+import Profile from "../user/profile/Profile";
+import Login from "../user/login/Login";
+import Signup from "../user/signup/Signup";
+import PrivateRoute from "../auth/PrivateRoute";
+import {useAuth} from "../auth/AuthProvider";
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      authenticated: false,
-      currentUser: null,
-      loading: true
+const App = () => {
+    const [loading, setLoading] = useState(true);
+    const {loadCurrentlyLoggedInUser} = useAuth();
+
+    const handleLogout = () => {
+        localStorage.removeItem(ACCESS_TOKEN);
+        // Alert.success("You're safely logged out!");
     }
 
-    this.loadCurrentlyLoggedInUser = this.loadCurrentlyLoggedInUser.bind(this);
-    this.handleLogout = this.handleLogout.bind(this);
-  }
+    useEffect(() => {
+        loadCurrentlyLoggedInUser(setLoading);
+    }, []);
 
-  loadCurrentlyLoggedInUser() {
-    getCurrentUser()
-    .then(response => {
-      this.setState({
-        currentUser: response,
-        authenticated: true,
-        loading: false
-      });
-    }).catch(error => {
-      this.setState({
-        loading: false
-      });  
-    });    
-  }
-
-  handleLogout() {
-    localStorage.removeItem(ACCESS_TOKEN);
-    this.setState({
-      authenticated: false,
-      currentUser: null
-    });
-    Alert.success("You're safely logged out!");
-  }
-
-  componentDidMount() {
-    this.loadCurrentlyLoggedInUser();
-  }
-
-  render() {
-    if(this.state.loading) {
-      return <LoadingIndicator />
+    if (loading) {
+        return <LoadingIndicator/>
     }
 
     return (
-      <div className="app">
-        <div className="app-top-box">
-          <AppHeader authenticated={this.state.authenticated} onLogout={this.handleLogout} />
+        <div className="app">
+            <BrowserRouter>
+                <div className="app-top-box">
+                    <AppHeader onLogout={handleLogout}/>
+                </div>
+                <div className="app-body">
+                    <Routes>
+                        <Route exact path="/" element={<Home/>}/>
+                        <Route path="/profile" element={<PrivateRoute><Profile/></PrivateRoute>}></Route>
+                        <Route path="/login" element={<Login/>}/>
+                        <Route path="/signup" element={<Signup/>}/>
+                        <Route path="/oauth2/redirect" element={<OAuth2RedirectHandler/>}/>
+                        <Route element={<NotFound/>}/>
+                    </Routes>
+                </div>
+            </BrowserRouter>
+            {/*<Alert stack={{limit: 3}}*/}
+            {/*       timeout={3000}*/}
+            {/*       position='top-right' effect='slide' offset={65}/>*/}
         </div>
-        <div className="app-body">
-          <Switch>
-            <Route exact path="/" component={Home}></Route>           
-            <PrivateRoute path="/profile" authenticated={this.state.authenticated} currentUser={this.state.currentUser}
-              component={Profile}></PrivateRoute>
-            <Route path="/login"
-              render={(props) => <Login authenticated={this.state.authenticated} {...props} />}></Route>
-            <Route path="/signup"
-              render={(props) => <Signup authenticated={this.state.authenticated} {...props} />}></Route>
-            <Route path="/oauth2/redirect" component={OAuth2RedirectHandler}></Route>  
-            <Route component={NotFound}></Route>
-          </Switch>
-        </div>
-        <Alert stack={{limit: 3}} 
-          timeout = {3000}
-          position='top-right' effect='slide' offset={65} />
-      </div>
     );
-  }
 }
 
 export default App;
