@@ -1,35 +1,27 @@
-import React, { Component } from 'react';
-import { ACCESS_TOKEN } from '../../constants';
-import { Redirect } from 'react-router-dom'
+import React from 'react';
+import {ACCESS_TOKEN} from '../../constants';
+import {Navigate} from "react-router-dom";
 
-class OAuth2RedirectHandler extends Component {
-    getUrlParameter(name) {
-        name = name.replace(/[\[]/, '\\[').replace(/[\]]/, '\\]');
-        var regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
+const getUrlParameter = (name, location) => {
+    name = name.replace(/[[]/, '\\[').replace(/[\]]/, '\\]');
+    const regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
+    const results = regex.exec(location.search);
+    return results === null ? '' : decodeURIComponent(results[1].replace(/\+/g, ' '));
+};
 
-        var results = regex.exec(this.props.location.search);
-        return results === null ? '' : decodeURIComponent(results[1].replace(/\+/g, ' '));
-    };
+const OAuth2RedirectHandler = (props) => {
 
-    render() {        
-        const token = this.getUrlParameter('token');
-        const error = this.getUrlParameter('error');
+    const token = getUrlParameter('token', props.location);
+    const error = getUrlParameter('error', props.location);
 
-        if(token) {
-            localStorage.setItem(ACCESS_TOKEN, token);
-            return <Redirect to={{
-                pathname: "/profile",
-                state: { from: this.props.location }
-            }}/>; 
-        } else {
-            return <Redirect to={{
-                pathname: "/login",
-                state: { 
-                    from: this.props.location,
-                    error: error 
-                }
-            }}/>; 
-        }
+    if (token) {
+        localStorage.setItem(ACCESS_TOKEN, token);
+        return <Navigate to="/" replace={true} state={{from: props.location}}/>;
+    } else {
+        return <Navigate to="/" replace={true} state={{
+            from: props.location,
+            error: error
+        }}/>;
     }
 }
 
