@@ -1,9 +1,11 @@
-import React from 'react';
+import React from "react";
 
-export default function LoadingIndicator(props) {
+const LoadingIndicator = () => {
     return (
         <div className="loading-indicator" style={{display: 'block', textAlign: 'center', marginTop: '30px'}}>
             Loading ...
         </div>
     );
 }
+
+export default LoadingIndicator;
