@@ -1,31 +1,23 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Home.css';
+import Circle from "./Circle";
 
-class Home extends Component {
-    render() {
-        return (
-            <div className="home-container">
-                <div className="container">
-                    <div className="graf-bg-container">
-                        <div className="graf-layout">
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                            <div className="graf-circle"></div>
-                        </div>
+const CIRCLE_AMOUNT = 12;
+
+const Home = () => {
+    return (
+        <div className="home-container">
+            <div className="container">
+                <div className="graf-bg-container">
+                    <div className="graf-layout">
+                        {Array.from({length: CIRCLE_AMOUNT}, (_, i) =>
+                            <Circle key={i}/>)}
                     </div>
-                    <h1 className="home-title">Spring Boot React OAuth2 Social Login Demo</h1>
                 </div>
+                <h1 className="home-title">Spring Boot React OAuth2 Social Login Demo</h1>
             </div>
-        )
-    }
+        </div>
+    )
 }
 
 export default Home;
