@@ -1,126 +1,49 @@
-import React, { Component } from 'react';
+import React, {useEffect} from 'react';
 import './Login.css';
-import { GOOGLE_AUTH_URL, FACEBOOK_AUTH_URL, GITHUB_AUTH_URL, ACCESS_TOKEN } from '../../constants';
-import { login } from '../../util/APIUtils';
-import { Link, Redirect } from 'react-router-dom'
-import fbLogo from '../../img/fb-logo.png';
-import googleLogo from '../../img/google-logo.png';
-import githubLogo from '../../img/github-logo.png';
-import Alert from 'react-s-alert';
+import {Navigate, useLocation, useNavigate} from 'react-router-dom'
+import SocialLogin from "./SocialLogin";
+import LoginForm from "./LoginForm";
+import {useAuth} from "../../auth/AuthProvider";
 
-class Login extends Component {
-    componentDidMount() {
+const Login = () => {
+    let {isAuthenticated} = useAuth();
+    let location = useLocation();
+    let navigate = useNavigate();
+
+    const redirectIfError = () => {
         // If the OAuth2 login encounters an error, the user is redirected to the /login page with an error.
         // Here we display the error and then remove the error query parameter from the location.
-        if(this.props.location.state && this.props.location.state.error) {
+        if (location.state && location.state.error) {
             setTimeout(() => {
-                Alert.error(this.props.location.state.error, {
-                    timeout: 5000
-                });
-                this.props.history.replace({
-                    pathname: this.props.location.pathname,
-                    state: {}
-                });
+                // Alert.error(location.state.error, {
+                //     timeout: 5000
+                // });
+                navigate(location.pathname, {replace: true, state: {}})
             }, 100);
         }
     }
-    
-    render() {
-        if(this.props.authenticated) {
-            return <Redirect
-                to={{
-                pathname: "/",
-                state: { from: this.props.location }
-            }}/>;            
-        }
 
-        return (
-            <div className="login-container">
-                <div className="login-content">
-                    <h1 className="login-title">Login to SpringSocial</h1>
-                    <SocialLogin />
-                    <div className="or-separator">
-                        <span className="or-text">OR</span>
-                    </div>
-                    <LoginForm {...this.props} />
-                    <span className="signup-link">New user? <Link to="/signup">Sign up!</Link></span>
+    useEffect(() => {
+        redirectIfError()
+    }, []);
+
+    if (isAuthenticated) {
+        return <Navigate to="/" replace={true} state={{from: location}}/>;
+    }
+    // <span className="signup-link">New user? <Link to="/signup">Sign up!</Link></span> - 47
+    return (
+        <div className="login-container">
+            <div className="login-content">
+                <h1 className="login-title">Login to SpringSocial</h1>
+                <SocialLogin/>
+                <div className="or-separator">
+                    <span className="or-text">OR</span>
                 </div>
+                <LoginForm />
+
             </div>
-        );
-    }
+        </div>
+    );
 }
 
-class SocialLogin extends Component {
-    render() {
-        return (
-            <div className="social-login">
-                <a className="btn btn-block social-btn google" href={GOOGLE_AUTH_URL}>
-                    <img src={googleLogo} alt="Google" /> Log in with Google</a>
-                <a className="btn btn-block social-btn facebook" href={FACEBOOK_AUTH_URL}>
-                    <img src={fbLogo} alt="Facebook" /> Log in with Facebook</a>
-                <a className="btn btn-block social-btn github" href={GITHUB_AUTH_URL}>
-                    <img src={githubLogo} alt="Github" /> Log in with Github</a>
-            </div>
-        );
-    }
-}
-
-
-class LoginForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            email: '',
-            password: ''
-        };
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    handleInputChange(event) {
-        const target = event.target;
-        const inputName = target.name;        
-        const inputValue = target.value;
-
-        this.setState({
-            [inputName] : inputValue
-        });        
-    }
-
-    handleSubmit(event) {
-        event.preventDefault();   
-
-        const loginRequest = Object.assign({}, this.state);
-
-        login(loginRequest)
-        .then(response => {
-            localStorage.setItem(ACCESS_TOKEN, response.accessToken);
-            Alert.success("You're successfully logged in!");
-            this.props.history.push("/");
-        }).catch(error => {
-            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');
-        });
-    }
-    
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <div className="form-item">
-                    <input type="email" name="email" 
-                        className="form-control" placeholder="Email"
-                        value={this.state.email} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <input type="password" name="password" 
-                        className="form-control" placeholder="Password"
-                        value={this.state.password} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <button type="submit" className="btn btn-block btn-primary">Login</button>
-                </div>
-            </form>                    
-        );
-    }
-}
-
-export default Login
+export default Login;
