@@ -1,36 +1,33 @@
-import React, { Component } from 'react';
+import React from "react";
 import './Profile.css';
+import {useAuth} from "../../auth/AuthProvider";
 
-class Profile extends Component {
-    constructor(props) {
-        super(props);
-        console.log(props);
-    }
-    render() {
-        return (
-            <div className="profile-container">
-                <div className="container">
-                    <div className="profile-info">
-                        <div className="profile-avatar">
-                            { 
-                                this.props.currentUser.imageUrl ? (
-                                    <img src={this.props.currentUser.imageUrl} alt={this.props.currentUser.name}/>
-                                ) : (
-                                    <div className="text-avatar">
-                                        <span>{this.props.currentUser.name && this.props.currentUser.name[0]}</span>
-                                    </div>
-                                )
-                            }
-                        </div>
-                        <div className="profile-name">
-                           <h2>{this.props.currentUser.name}</h2>
-                           <p className="profile-email">{this.props.currentUser.email}</p>
-                        </div>
+const Profile = () => {
+    const {currentUser} = useAuth();
+
+    return (
+        <div className="profile-container">
+            <div className="container">
+                <div className="profile-info">
+                    <div className="profile-avatar">
+                        {
+                            currentUser.imageUrl ? (
+                                <img src={currentUser.imageUrl} alt={currentUser.name}/>
+                            ) : (
+                                <div className="text-avatar">
+                                    <span>{currentUser.name && currentUser.name[0]}</span>
+                                </div>
+                            )
+                        }
                     </div>
-                </div>    
+                    <div className="profile-name">
+                        <h2>{currentUser.name}</h2>
+                        <p className="profile-email">{currentUser.email}</p>
+                    </div>
+                </div>
             </div>
-        );
-    }
+        </div>
+    );
 }
 
 export default Profile
