@@ -1,116 +1,31 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Signup.css';
-import { Link, Redirect } from 'react-router-dom'
-import { GOOGLE_AUTH_URL, FACEBOOK_AUTH_URL, GITHUB_AUTH_URL } from '../../constants';
-import { signup } from '../../util/APIUtils';
-import fbLogo from '../../img/fb-logo.png';
-import googleLogo from '../../img/google-logo.png';
-import githubLogo from '../../img/github-logo.png';
-import Alert from 'react-s-alert';
+import SignupForm from "./SignupForm";
+import SocialSignup from "./SocialSignup";
+import {Navigate, useLocation} from "react-router-dom";
+import {useAuth} from "../../auth/AuthProvider";
 
-class Signup extends Component {
-    render() {
-        if(this.props.authenticated) {
-            return <Redirect
-                to={{
-                pathname: "/",
-                state: { from: this.props.location }
-            }}/>;            
-        }
+const Signup = () => {
+    let {isAuthenticated} = useAuth();
+    let location = useLocation();
 
-        return (
-            <div className="signup-container">
-                <div className="signup-content">
-                    <h1 className="signup-title">Signup with SpringSocial</h1>
-                    <SocialSignup />
-                    <div className="or-separator">
-                        <span className="or-text">OR</span>
-                    </div>
-                    <SignupForm {...this.props} />
-                    <span className="login-link">Already have an account? <Link to="/login">Login!</Link></span>
+    if (isAuthenticated) {
+        return <Navigate to="/" replace state={{from: location}}/>;
+    }
+    // <span className="login-link">Already have an account? <Link to="/login">Login!</Link></span> - 25
+    return (
+        <div className="signup-container">
+            <div className="signup-content">
+                <h1 className="signup-title">Signup with SpringSocial</h1>
+                <SocialSignup/>
+                <div className="or-separator">
+                    <span className="or-text">OR</span>
                 </div>
+                <SignupForm />
+
             </div>
-        );
-    }
-}
-
-
-class SocialSignup extends Component {
-    render() {
-        return (
-            <div className="social-signup">
-                <a className="btn btn-block social-btn google" href={GOOGLE_AUTH_URL}>
-                    <img src={googleLogo} alt="Google" /> Sign up with Google</a>
-                <a className="btn btn-block social-btn facebook" href={FACEBOOK_AUTH_URL}>
-                    <img src={fbLogo} alt="Facebook" /> Sign up with Facebook</a>
-                <a className="btn btn-block social-btn github" href={GITHUB_AUTH_URL}>
-                    <img src={githubLogo} alt="Github" /> Sign up with Github</a>
-            </div>
-        );
-    }
-}
-
-class SignupForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            name: '',
-            email: '',
-            password: ''
-        }
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    handleInputChange(event) {
-        const target = event.target;
-        const inputName = target.name;        
-        const inputValue = target.value;
-
-        this.setState({
-            [inputName] : inputValue
-        });        
-    }
-
-    handleSubmit(event) {
-        event.preventDefault();   
-
-        const signUpRequest = Object.assign({}, this.state);
-
-        signup(signUpRequest)
-        .then(response => {
-            Alert.success("You're successfully registered. Please login to continue!");
-            this.props.history.push("/login");
-        }).catch(error => {
-            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');            
-        });
-    }
-
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <div className="form-item">
-                    <input type="text" name="name" 
-                        className="form-control" placeholder="Name"
-                        value={this.state.name} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <input type="email" name="email" 
-                        className="form-control" placeholder="Email"
-                        value={this.state.email} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <input type="password" name="password" 
-                        className="form-control" placeholder="Password"
-                        value={this.state.password} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <button type="submit" className="btn btn-block btn-primary" >Sign Up</button>
-                </div>
-            </form>                    
-
-        );
-    }
+        </div>
+    );
 }
 
 export default Signup
