@@ -1,8 +1,8 @@
 package com.example.springsocial.model;
 
 public enum  AuthProvider {
-    local,
-    facebook,
-    google,
-    github
+    LOCAL,
+    FACEBOOK,
+    GOOGLE,
+    GITHUB
 }
